State synchronization
In effect, this fixes most of the UI issues where the state is not
updated (e.g. data usage, test completion, etc.)

This also moves the state management and test running to the background
page, which allows the popup to render as a view of the internal state.
This fixes the issue where the popup "loses" its place if a test is
in progress. (Settings are still saved via the popup, and then the
background page is notified to update schedules)

Simplified the measure controller and how the guage intervals were
being rendered: it should look more like 50% up, 50% down, then
complete, and shows the "More" button to add annotation to the latest
test.

Data visualization has a simple statistic toggle for "Upload" vs.
"Download".

Lots of refactoring overall, simplifying state to use StorageService and
promises to deliver the retrieved data back through the
`ChromeAppSupport.notify` message bus, which are either resolved
directly or, in the case of most controllers, via `$rootScope.$emit`
listeners.