Fix CSV export.
CSV export used to fail in presence of a mix of old and new measurements in the
history. This fix makes sure the export works again by normalizing field names
(using the newest ones) and adding an explicit "version" field on history
records so that they can easily be distinguished.